requirements.txt
astronomer-cosmos
pandas
sqlalchemy
requests
apache-airflow-providers-postgres
pyarrow
psycopg2-binary